+++ pyVoIP/regex.py
@@ -19,3 +19,4 @@ TO_FROM_DIRECT_MATCH = re.compile(
     r'(?P<display_name>"?[\w ]*"? )?<?(?P<uri_type>sips?):(?P<host>[\w.]+)(?P<port>:[0-9]+)?>?'
 )
 SDP_A_SPLIT = re.compile(" |/")
+SIP_VERSION_MATCH = re.compile(r"(?:SIP|sip)/[0-9.]+")
